Bar: Adaugarea seed

+++ nightClub.Domain/Entities/Bar/BarDbTable.cs
@@ -1,10 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations.Schema;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using System.ComponentModel.DataAnnotations.Schema;
 
 namespace nightClub.Domain.Entities.Bar
 {
@@ -16,8 +12,8 @@ namespace nightClub.Domain.Entities.Bar
         public string Url { get; set; }
         public string Title { get; set; }
         public string Category { get; set; }
-        public decimal Alcohol { get; set; }
-        public decimal Price { get; set; }
+        public double Alcohol { get; set; }
+        public double Price { get; set; }
         public string Description { get; set; }
         public DateTime Date { get; set; }
     }
