+++ nightClub.Domain/Entities/Bar/PhotoBar.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace nightClub.Domain.Entities.Bar
 {
@@ -12,8 +8,8 @@ namespace nightClub.Domain.Entities.Bar
         public string Url { get; set; }
         public string Title { get; set; }
         public string Category { get; set; }
-        public decimal Alcohol { get; set; }
-        public decimal Price { get; set; }
+        public double Alcohol { get; set; }
+        public double Price { get; set; }
         public string Description { get; set; }
         public DateTime Date { get; set; }
     }
